refactor: altera arquivo que fará a requisição
