+++ src/routes/Aquarium/Info/index.tsx
@@ -1,43 +1,13 @@
-import React from 'react'
 import { useMedia } from '@/hooks'
 import { Box } from '@chakra-ui/react'
-import { useParams } from 'react-router-dom'
-import { AppDispatch } from '@/redux/store'
-import { useDispatch, useSelector } from 'react-redux'
-import { AquariumDetailsTopBar, DetailsBox } from '@/components'
-import {
-  fetchAquariumDetails,
-  getAquariumDetails,
-  resetAquarium,
-} from '@/redux/reducers/aquariumDetails'
-import {
-  fetchPetsByAquariumId,
-  getPetsByAquariumId,
-  resetPets,
-} from '@/redux/reducers/petsDetails'
+import { useSelector } from 'react-redux'
 import { SensorDetails } from '@/components/SensorDetails'
-import {
-  fetchAccessoriesAquarium,
-  resetAccessories,
-} from '@/redux/reducers/accessoriesDetails'
+import { AquariumDetailsTopBar, DetailsBox } from '@/components'
+import { getAquariumDetails } from '@/redux/reducers/aquariumDetails'
+import { getPetsByAquariumId } from '@/redux/reducers/petsDetails'
 
 export const AquariumInfo = () => {
   const { isDesktop } = useMedia()
-  const { id } = useParams()
-  const dispatch = useDispatch<AppDispatch>()
-
-  const fetchAquarium = React.useCallback(async () => {
-    dispatch(fetchAquariumDetails(id!))
-    dispatch(fetchPetsByAquariumId(id!))
-    dispatch(fetchAccessoriesAquarium(id!))
-  }, [dispatch, id])
-
-  React.useEffect(() => {
-    dispatch(resetPets())
-    dispatch(resetAquarium())
-    dispatch(resetAccessories())
-    fetchAquarium()
-  }, [dispatch, fetchAquarium])
 
   const details = useSelector(getAquariumDetails)
   const pets = useSelector(getPetsByAquariumId)
